feat: ini auth
