+++ apps/app/pages/api/auth/[...nextauth].ts
@@ -13,81 +13,23 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import NextAuth from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { getCsrfToken } from "next-auth/react";
-import { SiweMessage } from "siwe";
-import { prisma } from "@lightdotso/prisma";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import type { NextApiRequest, NextApiResponse } from "next";
+import NextAuth, { type NextAuthOptions } from "next-auth";
+import { authOptions } from "@lightdotso/auth";
 
-// From: https://github.com/spruceid/siwe-next-auth-example
-// For more information on each option (and a full list of options) go to
-// https://next-auth.js.org/configuration/options
-export default async function auth(req: any, res: any) {
-  const providers = [
-    CredentialsProvider({
-      name: "Ethereum",
-      credentials: {
-        message: {
-          label: "Message",
-          type: "text",
-          placeholder: "0x0",
-        },
-        signature: {
-          label: "Signature",
-          type: "text",
-          placeholder: "0x0",
-        },
-      },
-      async authorize(credentials) {
-        try {
-          const siwe = new SiweMessage(
-            JSON.parse(credentials?.message || "{}"),
-          );
-          const nextAuthUrl = new URL(process.env.NEXTAUTH_URL!);
-
-          const result = await siwe.verify({
-            signature: credentials?.signature || "",
-            domain: nextAuthUrl.host,
-            nonce: await getCsrfToken({ req: { headers: req.headers } }),
-          });
-
-          if (result.success) {
-            return {
-              id: siwe.address,
-            };
-          }
-          return null;
-        } catch (e) {
-          return null;
-        }
-      },
-    }),
-  ];
+const Auth = (req: NextApiRequest, res: NextApiResponse) => {
+  const authOpts: NextAuthOptions = authOptions({ req });
 
   const isDefaultSigninPage =
-    req.method === "GET" && req.query.nextauth.includes("signin");
+    req.method === "GET" && req?.query?.nextauth?.includes("signin");
 
   // Hide Sign-In with Ethereum from default sign page
   if (isDefaultSigninPage) {
-    providers.pop();
+    // Removes from the authOptions.providers array
+    authOpts.providers.pop();
   }
 
-  return await NextAuth(req, res, {
-    // https://next-auth.js.org/configuration/providers/oauth
-    providers,
-    session: {
-      strategy: "jwt",
-    },
-    adapter: PrismaAdapter(prisma),
-    secret: process.env.NEXTAUTH_SECRET!,
-    callbacks: {
-      async session({ session, token }: { session: any; token: any }) {
-        session.address = token.sub;
-        session.user.name = token.sub;
-        session.user.image = "/";
-        return session;
-      },
-    },
-  });
-}
+  return NextAuth(req, res, authOpts) as typeof NextAuth;
+};
+
+export default Auth;
