+++ packages/auth/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lightdotso/trpc",
+  "name": "@lightdotso/auth",
   "version": "0.0.0",
   "private": false,
   "homepage": "https://github.com/LightDotSo/LightDotSo",
@@ -14,23 +14,11 @@
   "author": "Light, Inc.",
   "main": "index.ts",
   "dependencies": {
-    "@lightdotso/auth": "workspace:*",
     "@lightdotso/prisma": "workspace:*",
-    "@prisma/client": "^5.0.0",
-    "@trpc-limiter/core": "^0.0.8",
-    "@trpc/server": "^10.35.0",
-    "@upstash/ratelimit": "^0.4.3",
-    "@upstash/redis": "^1.22.0",
+    "@next-auth/prisma-adapter": "0.4.4-canary.81",
     "next": "^13.4.12",
     "next-auth": "^4.22.3",
-    "openapi-types": "^12.1.3",
-    "prisma": "^5.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "superjson": "^1.13.1",
-    "swagger-ui-react": "^5.2.0",
-    "trpc-openapi": "^1.2.0",
-    "zod": "^3.21.4"
+    "siwe": "^2.1.4"
   },
   "devDependencies": {
     "@lightdotso/tsconfig": "workspace:*",
