+++ packages/auth/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "outDir": "./dist"
+  },
+  "exclude": ["dist", "build", "node_modules"],
+  "extends": "@lightdotso/tsconfig/node.json",
+  "include": ["."]
+}
